+++ apps/native-component-list/src/screens/UI/ButtonScreen.ios.tsx
@@ -1,11 +1,13 @@
 import {
-  Button as ButtonPrimitive,
+  Button,
   CircularProgress,
   Host,
   Image,
   Label,
+  List,
   Text,
   VStack,
+  Section,
 } from '@expo/ui/swift-ui';
 import {
   background,
@@ -15,16 +17,13 @@ import {
   shapes,
 } from '@expo/ui/swift-ui/modifiers';
 import * as React from 'react';
-import { ScrollView, StyleProp, StyleSheet, ViewStyle } from 'react-native';
-
-import { Page, Section } from '../../components/Page';
 
 export default function ButtonScreen() {
   return (
-    <Page>
-      <ScrollView showsVerticalScrollIndicator={false}>
+    <Host style={{ flex: 1 }}>
+      <List>
         <Section title="With Label">
-          <Button style={styles.button}>
+          <Button>
             <Label
               title="Default Padding"
               systemImage="gear"
@@ -35,14 +34,14 @@ export default function ButtonScreen() {
               ]}
             />
           </Button>
-          <Button style={styles.button}>
+          <Button>
             <Label
               title="No Padding"
               systemImage="gear"
               modifiers={[foregroundStyle('black'), padding({ all: 0 }), background('yellow')]}
             />
           </Button>
-          <Button style={styles.button}>
+          <Button>
             <Label
               title="Custom Padding"
               modifiers={[
@@ -54,151 +53,82 @@ export default function ButtonScreen() {
           </Button>
         </Section>
         <Section title="Default">
-          <Button style={styles.button}>Test</Button>
+          <Button>Test</Button>
         </Section>
         <Section title="System Styles">
-          <Button style={styles.button} variant="default">
-            Default
-          </Button>
-          <Button style={styles.button} variant="glass">
-            Glass button
-          </Button>
-          <Button style={styles.button} variant="glassProminent">
-            Glass Prominent
-          </Button>
-          <Button style={styles.button} variant="bordered">
-            Bordered
-          </Button>
-          <Button style={styles.button} variant="borderless">
-            Borderless
-          </Button>
-          <Button style={styles.button} variant="borderedProminent">
-            Bordered Prominent
-          </Button>
-          <Button style={styles.button} variant="plain">
-            Plain
-          </Button>
+          <Button variant="default">Default</Button>
+          <Button variant="glass">Glass button</Button>
+          <Button variant="glassProminent">Glass Prominent</Button>
+          <Button variant="bordered">Bordered</Button>
+          <Button variant="borderless">Borderless</Button>
+          <Button variant="borderedProminent">Bordered Prominent</Button>
+          <Button variant="plain">Plain</Button>
         </Section>
         <Section title="Control Size">
-          <Button
-            style={styles.button}
-            controlSize="mini"
-            variant="glassProminent"
-            modifiers={[fixedSize()]}>
+          <Button controlSize="mini" variant="glassProminent" modifiers={[fixedSize()]}>
             Mini glass prominent
           </Button>
-          <Button style={styles.button} controlSize="small" variant="bordered">
+          <Button controlSize="small" variant="bordered">
             Small bordered
           </Button>
-          <Button style={styles.button} controlSize="regular" variant="glass">
+          <Button controlSize="regular" variant="glass">
             Regular glass
           </Button>
-          <Button style={styles.button} controlSize="large" variant="glassProminent">
+          <Button controlSize="large" variant="glassProminent">
             Large
           </Button>
-          <Button style={styles.button} controlSize="large" variant="glass">
+          <Button controlSize="large" variant="glass">
             Large glass
           </Button>
           <Button
-            style={styles.button}
             controlSize="extraLarge"
             variant="glassProminent"
             systemImage="square.and.arrow.up"
-            color="orange"
-            modifiers={[fixedSize()]}>
+            color="orange">
             Extra Large (iOS 17+)
           </Button>
         </Section>
         <Section title="Disabled">
-          <Button style={styles.button} disabled>
-            Disabled
-          </Button>
-          <Button style={styles.button}>Enabled</Button>
+          <Button disabled>Disabled</Button>
+          <Button>Enabled</Button>
         </Section>
         <Section title="Button Roles">
-          <Button style={styles.button} role="default">
-            Default
-          </Button>
-          <Button style={styles.button} role="cancel">
-            Cancel
-          </Button>
-          <Button style={styles.button} role="destructive">
-            Destructive
-          </Button>
+          <Button role="default">Default</Button>
+          <Button role="cancel">Cancel</Button>
+          <Button role="destructive">Destructive</Button>
         </Section>
         <Section title="Button Images">
-          <Button variant="bordered" style={styles.button} systemImage="folder">
+          <Button variant="bordered" systemImage="folder">
             Folder
           </Button>
-          <Button style={styles.button} systemImage="tortoise">
-            Tortoise
-          </Button>
-          <Button variant="borderless" style={styles.button} systemImage="trash">
+          <Button systemImage="tortoise">Tortoise</Button>
+          <Button variant="borderless" systemImage="trash">
             Trash
           </Button>
-          <Button style={styles.button} systemImage="heart">
-            Heart
-          </Button>
-          <Button style={styles.button} systemImage="gear" variant="glass" />
+          <Button systemImage="heart">Heart</Button>
+          <Button systemImage="gear" variant="glass" />
         </Section>
         <Section title="Tinted Buttons">
-          <Button style={styles.button} color="#f00f0f">
-            Red
-          </Button>
+          <Button color="#f00f0f">Red</Button>
         </Section>
         <Section title="Custom children">
-          <Host style={styles.buttonHost}>
-            <ButtonPrimitive>
-              <VStack spacing={4}>
-                <Image systemName="folder" />
-                <Text>Folder</Text>
-              </VStack>
-            </ButtonPrimitive>
-          </Host>
-          <Host style={styles.buttonHost}>
-            <ButtonPrimitive>
-              <CircularProgress color="blue" />
-            </ButtonPrimitive>
-          </Host>
+          <Button>
+            <VStack spacing={4}>
+              <Image systemName="folder" />
+              <Text>Folder</Text>
+            </VStack>
+          </Button>
+          <Button>
+            <CircularProgress color="blue" />
+          </Button>
         </Section>
         <Section title="interpolated strings">
-          <Button style={styles.button} color="#FF6347">
+          <Button color="#FF6347">
             {/* eslint-disable-next-line */}
             Hello {'world'}
           </Button>
         </Section>
-      </ScrollView>
-    </Page>
-  );
-}
-
-function Button(
-  props: React.ComponentProps<typeof ButtonPrimitive> & { style?: StyleProp<ViewStyle> }
-) {
-  const { style, ...restProps } = props;
-  return (
-    <Host matchContents style={style}>
-      <ButtonPrimitive {...restProps}>{props.children}</ButtonPrimitive>
+      </List>
     </Host>
   );
 }
-
-const styles = StyleSheet.create({
-  button: {
-    width: 150,
-    margin: 5,
-    marginLeft: 20,
-    overflow: 'visible',
-  },
-  buttonHost: {
-    width: 50,
-    height: 50,
-  },
-  stretch: {
-    alignSelf: 'stretch',
-  },
-  columnWrapper: {
-    justifyContent: 'space-around',
-    alignContent: 'space-around',
-  },
-});