+++ packages/expo-ui/ios/Label.swift
@@ -6,7 +6,6 @@ import SwiftUI
 final class LabelViewProps: UIBaseViewProps {
   @Field var title: String?
   @Field var systemImage: String?
-  @Field var color: Color?
 }
 
 struct LabelView: ExpoSwiftUI.View {
@@ -14,19 +13,10 @@ struct LabelView: ExpoSwiftUI.View {
 
   var body: some View {
     if let title = props.title, let systemImage = props.systemImage {
-      // TODO: Deprecate this - recommend using foregroundStyle modifier
-      if let color = props.color {
-        Label(title, systemImage: systemImage).foregroundStyle(color)
-          .applyFixedSize(props.fixedSize ?? true)
-      } else {
-        Label(title, systemImage: systemImage)
-          .applyFixedSize(props.fixedSize ?? true)
-      }
+      Label(title, systemImage: systemImage)
     }
-    // TODO: Deprecate this - recommend using labelStyle modifier
     else if let title = props.title {
       Label(title, systemImage: "").labelStyle(.titleOnly)
-        .applyFixedSize(props.fixedSize ?? true)
     }
   }
 }