+++ packages/expo-ui/ios/UIBaseView.swift
@@ -15,9 +15,6 @@ public struct UIBaseView<Props: UIBaseViewProps, Content: ExpoSwiftUI.View<Props
 
   public var body: some View {
     Content(props: props)
-      .applyFixedSize(props.fixedSize)
-      .applyFrame(props.frame, defaultAlignment: props.defaultFrameAlignment)
-      .applyPadding(props.padding)
       .applyAccessibilityIdentifier(props.testID)
       .applyModifiers(props.modifiers, appContext: props.appContext, globalEventDispatcher: props.globalEventDispatcher)
   }
@@ -33,9 +30,6 @@ public struct UIBaseViewModifier<Props: UIBaseViewProps>: ViewModifier {
 
   public func body(content: Content) -> some View {
     content
-      .applyFixedSize(props.fixedSize)
-      .applyFrame(props.frame, defaultAlignment: props.defaultFrameAlignment)
-      .applyPadding(props.padding)
       .applyAccessibilityIdentifier(props.testID)
       .applyModifiers(props.modifiers, appContext: props.appContext, globalEventDispatcher: props.globalEventDispatcher)
   }